Made the pycllp solver Scenario aware.
In order to do this new methods get_all_* methods are added the Node and Storage objects. These provide convenient access to all the scenario data in one call. This means the solvers do not need to iterate through all scenario combinations, and can modify and solve them all simultaneously.